Update to GTFS 4.x
Possible thanks to MagicMirror updating to a new node.js major
version! (and better JS understanding on my part!)

This updates to the newest version of gtfs, which simplifies realtime
monitoring logic and seems to load much faster.

There are a few breaking changes that'll require updates to config.js.